Refactor file attachment handlers into custom hook
Extract duplicate file attachment logic from TextGenerationChatContainer
and TextGenerationCompletionContainer into a reusable useFileAttachments
hook. Also adds MAX_IMAGE_FILE_SIZE constant usage.
